backend/requirements/base.txt
Django==1.8.2
django-celery==3.1.17
  celery==3.1.23
    billiard==3.3.0.23
    kombu==3.0.34
      amqp==1.4.9
      anyjson==0.3.3
django-celery-transactions==0.3.5
Babel==2.3.4
BeautifulSoup==3.2.1
data-importer==3.0.2
dj-database-url==0.4.1
django-autoslug==1.9.3
-e git+https://github.com/yourlabs/django-cities-light.git@3bc210cff6714084be4a8ff6b4d640209bcdaa6e#egg=django_cities_light
django-classy-tags==0.8.0
django-configurations==2.0
django-environ==0.4.1
-e git+https://github.com/divio/django-filer.git@76131197bfee87824599aaac8624ad6ce4e32fbe#egg=django_filer
django-mptt==0.8.7
django-paging==0.2.5
django-phonenumber-field==1.1.0
django-polymorphic==0.8.1
django-redis==4.7.0
django-registration==2.2
django-sekizai==0.10.0
django-smart-selects==1.2.6
django-split-settings==0.2.4
django-tastypie==0.13.3
django-templatetag-sugar==1.0
django-test-utils==0.3
easy-thumbnails==2.3
et-xmlfile==1.0.1
gunicorn==19.3.0
jdcal==1.3
mimeparse==0.1.3
MySQL-python==1.2.5
openpyxl==2.4.0
phonenumberslite==8.0.0
Pillow==3.4.2
pipestat==0.2.0
progressbar2==3.12.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
python-utils==2.0.1
pytz==2016.10
redis==2.10.5
six==1.10.0
sqlparse==0.2.2
Unidecode==0.4.19
vine==1.1.3
xlrd==1.0.0
